Ajuste ID Profissional de Integer para Long

+++ src/main/java/mierim/model/repository/ProfissionalRepository.java
@@ -4,5 +4,5 @@ import mierim.model.entity.Cidade;
 import mierim.model.entity.Profissional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfissionalRepository  extends JpaRepository<Profissional, Integer> {
+public interface ProfissionalRepository  extends JpaRepository<Profissional, Long> {
 }
